Fix an issue of selecting sub_fields
When selecting a sub_field,
sometimes it also selected
a non-sub_field with the same name,
like 'new-field-1'
